+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.1] - unreleased
+
+### Added
+
+- Option to disable query runner
+- Progress indicator when requesting metadata from profile
+- Option to filter metadata based on XPath expressions
+- Default XPath filter for Annotations
+
+### Fixed
+
+- Model dependent token limits instead of hardcoded limits
+- Request metadata from treeview fixed
+- Trim profile endpoint URL
+
+### Removed
+
+- Settings `odata.removeAnnotations` removed, this is now default behavior and implemented via XPath
+
 ## [0.2.0] - 2025-04-13
 
 ### Added